Refactor: Move the functions about `complete_default` to a new file `read_pp_complete.cpp`.
