+++ source/Makefile.Objects
@@ -171,6 +171,7 @@ OBJS_CELL=atom_pseudo.o\
     atom_spec.o\
     pseudo.o\
     read_pp.o\
+    read_pp_complete.o\
     read_pp_upf100.o\
     read_pp_upf201.o\
     read_pp_blps.o\
