+++ source/module_cell/test_pw/CMakeLists.txt
@@ -12,7 +12,7 @@ AddTest(
   TARGET cell_unitcell_test_pw
   LIBS ${math_libs} base device parameter
   SOURCES unitcell_test_pw.cpp ../unitcell.cpp ../read_atoms.cpp ../read_cell_pseudopots.cpp ../atom_spec.cpp
-	../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp
+	../atom_pseudo.cpp ../pseudo.cpp ../read_pp.cpp ../read_pp_complete.cpp ../read_pp_upf201.cpp ../read_pp_upf100.cpp
 	../read_pp_vwr.cpp ../read_pp_blps.cpp ../../module_io/output.cpp
 )
